# api/cmd/api/docs/swagger.yaml
definitions:
  github_com_Caqil_megapdf-api_internal_models.Account:
    properties:
      accessToken:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      expiresAt:
        type: integer
      id:
        type: integer
      idToken:
        type: string
      provider:
        type: string
      providerAccountId:
        type: string
      refreshToken:
        type: string
      scope:
        type: string
      sessionState:
        type: string
      tokenType:
        type: string
      type:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.User'
      userId:
        type: string
    type: object
  github_com_Caqil_megapdf-api_internal_models.ApiKey:
    properties:
      createdAt:
        type: string
      expiresAt:
        type: string
      id:
        type: string
      key:
        type: string
      lastUsed:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.User'
      userId:
        type: string
    type: object
...
  github_com_Caqil_megapdf-api_internal_models.Session:
    properties:
      expires:
        type: string
      id:
        type: string
      sessionToken:
        type: string
      user:
        $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.User'
      userId:
        type: string
    type: object
...
  github_com_Caqil_megapdf-api_internal_models.Transaction:
    properties:
      amount:
        type: number
      balanceAfter:
        type: number
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      paymentId:
        type: string
      status:
        type: string
      user:
        $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.User'
      userId:
        type: string
    type: object
  github_com_Caqil_megapdf-api_internal_models.UsageStats:
    properties:
      count:
        type: integer
      createdAt:
        type: string
      date:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      operation:
        type: string
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.User'
      userId:
        type: string
    type: object
  github_com_Caqil_megapdf-api_internal_models.User:
    properties:
      accounts:
        items:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Account'
        type: array
      apiKeys:
        items:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.ApiKey'
        type: array
      balance:
        type: number
      createdAt:
        type: string
      email:
        type: string
      emailVerified:
        type: string
      freeOperationsReset:
        type: string
      freeOperationsUsed:
        type: integer
      id:
        type: string
      image:
        type: string
      isEmailVerified:
        type: boolean
      name:
        type: string
      password:
        type: string
      role:
        type: string
      sessions:
        items:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Session'
        type: array
      transactions:
        items:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Transaction'
        type: array
      updatedAt:
        type: string
      usageStats:
        items:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.UsageStats'
        type: array
      verificationToken:
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /sessions:
    post:
      consumes:
      - application/json
      description: Creates a new session with the provided details
      parameters:
      - description: Session data
        in: body
        name: session
        required: true
        schema:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Session'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Session'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new session
      tags:
      - sessions
  /sessions/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a session by its ID
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a session
      tags:
      - sessions
    get:
      consumes:
      - application/json
      description: Retrieves a session by its ID
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Session'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a session
      tags:
      - sessions
    put:
      consumes:
      - application/json
      description: Updates a session with the provided details
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Session data
        in: body
        name: session
        required: true
        schema:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Session'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Session'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a session
      tags:
      - sessions
  /transactions:
    post:
      consumes:
      - application/json
      description: Creates a new transaction with the provided details
      parameters:
      - description: Transaction data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Transaction'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Transaction'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new transaction
      tags:
      - transactions
  /transactions/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a transaction by its ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a transaction
      tags:
      - transactions
    get:
      consumes:
      - application/json
      description: Retrieves a transaction by its ID
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Transaction'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a transaction
      tags:
      - transactions
    put:
      consumes:
      - application/json
      description: Updates a transaction with the provided details
      parameters:
      - description: Transaction ID
        in: path
        name: id
        required: true
        type: string
      - description: Transaction data
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Transaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.Transaction'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a transaction
      tags:
      - transactions
  /usage-stats:
    post:
      consumes:
      - application/json
      description: Creates new usage stats with the provided details
      parameters:
      - description: UsageStats data
        in: body
        name: usageStats
        required: true
        schema:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.UsageStats'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.UsageStats'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create new usage stats
      tags:
      - usage-stats
  /usage-stats/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes usage stats by its ID
      parameters:
      - description: UsageStats ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete usage stats
      tags:
      - usage-stats
    get:
      consumes:
      - application/json
      description: Retrieves usage stats by its ID
      parameters:
      - description: UsageStats ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.UsageStats'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get usage stats
      tags:
      - usage-stats
    put:
      consumes:
      - application/json
      description: Updates usage stats with the provided details
      parameters:
      - description: UsageStats ID
        in: path
        name: id
        required: true
        type: string
      - description: UsageStats data
        in: body
        name: usageStats
        required: true
        schema:
          $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.UsageStats'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Caqil_megapdf-api_internal_models.UsageStats'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update usage stats
      tags:
      - usage-stats
...
swagger: "2.0"